src/requirement.txt
python-dotenv
duckduckgo-search
fastapi
uvicorn
groq
gunicorn